requirements.txt
asgiref==3.8.1
bleach==6.2.0
Django==5.2.3
django-ckeditor==6.7.3
django-ckeditor-5==0.2.18
django-js-asset==3.1.2
django-prose==2.1.0
django-summernote==0.8.20.0
django-tinymce==4.1.0
passlib==1.7.4
pillow==11.2.1
sqlparse==0.5.3
webencodings==0.5.1
